ELI-FILES/requirements.txt
# Python package requirements
discord.py
python-dotenv